Broker/requirements.txt
flask
flask_restx
colorlog
confluent-kafka
requests
psycopg2-binary